Disable VH check for the thinner floors of PPD12
In the stairway, a thinner floor is used then in the rest of the house but the assumption in the modelling are cubic zones (fixed height for entire zone) while reality the hallwayy will have two different heights. The choice was made to select the reference heights based on the thicker floor hence needing to disable the VH check for the thinner internal floors.